==> code/0_packages.R <==
library(readxl)
library(writexl)
library(faux)
library(ggplot2)
library(ggpubr)
library(dplyr)
library(purrr)
library(pROC) # roc(), roc.test()
library(tidyr)
library(stringr)
library(lmtest) # for lrtest()
library(gridExtra)
library(gtable)
library(ggtext)
library(cowplot) # as_grob()
library(grid)
library(sjPlot)
library(rempsyc)
library(stargazer)
library(tinytex)
library(knitr)
library(broom)
library(devtools)
library(kableExtra)
library(ggarchery)
